repeated code moved to services

+++ src/game/game.service.ts
@@ -53,11 +53,6 @@ export class GameService {
         return await this.gameRepository.save(newGame);
     }
 
-    async createGameSupply(gameSupply: CreateGameSupplyDto[]) {
-        const newItem = this.gameSupplyRepository.create(gameSupply);
-        return await this.gameSupplyRepository.save(newItem);
-    }
-
     async createGameNavigation(gameNavigation: CreateGameNavigationDto) {
         const newItem = this.gameNavigationRepository.create(gameNavigation);
         return await this.gameNavigationRepository.save(newItem);
@@ -65,61 +60,29 @@ export class GameService {
 
     async createCharacterQueue(characterQueue: CreateCharacterQueueDto[]) {
         const newItem = this.characterQueueRepository.create(characterQueue);
-        console.log(newItem);
         return await this.characterQueueRepository.save(newItem);
-    }
-
-    async drawSupplies(amount: number, gameId: number) {
-        const selected = await this.gameSupplyRepository
-        .createQueryBuilder("game_supply")
-        .where("game_supply.picked is false AND game_supply.gameId = :gameId", {gameId: gameId})
-        .orderBy("RANDOM()")
-        .limit(amount)
-        .getMany();
-        selected.forEach(supply => supply.picked = true);
-        return await this.gameSupplyRepository.save(selected)
-    }
-    async getDrawnSupplies(gameId: number) {
-        return await this.gameSupplyRepository
-        .createQueryBuilder("game_supply")
-        .leftJoinAndSelect("game_supply.supply", "supply")
-        .where("game_supply.picked AND game_supply.gameId = :gameId", {gameId: gameId})
-        .getMany();
     }
 
     async pickSupply(supply: string, player: Player, game: Game) {
         if(game.state != GameState.Supplies)
             throw new WsException('Action is unavailable during this phase');
-
         const currentCharacter = game.queue.find(character => character.order == game.currentCharacterIndex);
-        console.log(currentCharacter);
-        console.log(player.character);
         if(currentCharacter.characterName != player.character.name)
             throw new WsException('Not your turn');
-        const gameSupply = await this.gameSupplyRepository.findOne({
-            where: {
-                supplyName: supply,
-                gameId: game.id,
-                picked: true
-            },
-            relations: {supply: true},
-        });
-        console.log('deleting...');
-        console.log(gameSupply.id)
-        const res = await this.gameSupplyRepository.delete({id: gameSupply.id});
-        console.log('deleted', res);
+        const gameSupply = await this.cardsService.getDrawnSupply(game.id, supply);
+        if(gameSupply == null)
+            throw new WsException('Supply is not available to pick.');
+        await this.gameSupplyRepository.delete({id: gameSupply.id});
         player.closedCards.push(gameSupply.supply);
         await this.userService.updatePlayer(player);
     }
     async gameTurn(game: Game) {
-        console.log('2');
         let newIndex = game.currentCharacterIndex + 1;
         let newState = game.state;
         if(game.currentCharacterIndex == game.players.length - 1){
             newState = game.state == 1 ? 2 : 1;
             newIndex = 0;
         } 
-        console.log(3);
         await this.gameRepository.update({id: game.id}, {
             currentCharacterIndex: newIndex,
             state: newState
@@ -160,7 +123,7 @@ export class GameService {
         await this.userService.createPlayers(players);
         await this.createCharacterQueue(queue);
         await Promise.all(supplies.map( async (supply) => {
-            await this.createGameSupply(Array(supply.amount).fill({
+            await this.cardsService.createGameSupply(Array(supply.amount).fill({
                 gameId: game.id,
                 supplyName: supply.name
             }));
